Reload stylesheets by updating their query string
The previous method caused an annoying visual flicker, due to the
complete removal of the stylesheet, followed by its immediate
re-addition.

By updating the stylesheet's href attribute with a meaningless query
string parameter instead, we retain the existing styles up until the
stylesheet has reloaded and parsed fully, so that only the changes are
noticeable as a change.